paper_agent_project/paper_agent_project/requirements.txt
python-dotenv>=1.0.0
requests>=2.31.0
pypdf>=4.0.0
python-docx>=1.1.0
scikit-learn>=1.3.0
numpy>=1.24.0
scipy>=1.10.0
rich>=13.0.0
